requirements.txt
aiohttp==3.6.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
multidict==4.7.4
packaging==19.2
pluggy==0.13.1
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.5
pytest==5.3.2
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-rerunfailures==8.0
PyYAML==5.2
requests==2.22.0
six==1.13.0
slack==0.0.2
slackclient==2.5.0
urllib3==1.25.7
wcwidth==0.1.7
xlrd==1.2.0
yarl==1.4.2
zipp==0.6.0
